Extract a GPIO class and allow any number of pins
Making this code a bit more generic will be more useful if it was published as a
gem, some stepper motors use a different number of pins so it makes sense to
allow any number.

It would be nice to add half stepping to the stepper class but because I
personally aren't using it, it can wait until later.

Also, it may be handy to have a wrapper for the GPIO pins for uses other than
stepper motors.